Nourhan

+++ app/Services/Favorite/FavoriteService.php
@@ -2,13 +2,12 @@
 
 namespace App\Services\Favorite;
 
-use auth;
+use Exception;
 use App\Models\User\User;
 use Illuminate\Http\Request;
 use App\Models\Product\Product;
 use App\Traits\CacheManagerTrait;
 use Illuminate\Database\Eloquent\Collection;
-
 
 class FavoriteService
 {
@@ -24,7 +23,10 @@ class FavoriteService
     public function storeFavorite(Product $product)
     {
         $user = User::findOrFail(auth()->user()->id);
-        $user->favoriteProducts()->attach($product->id);
+        $result = $user->favoriteProducts()->syncWithoutDetaching($product->id);
+        if (empty($result['attached'])) {
+            throw new Exception("Product is already in favorites");
+        }
         $this->clearCacheGroup($this->groupe_key_cache);
     }
 
@@ -37,7 +39,11 @@ class FavoriteService
     public function showFavorites()
     {
         $user = User::findOrFail(auth()->id());
-        return  $user->favoriteProducts()->get();
+        $user_favorite_products = $user->favoriteProducts()->get();
+        if (empty($user_favorite_products)) {
+            throw new Exception("You do not have favorite  products,Add some product to favorite.", 404);
+        }
+        return $user_favorite_products;
     }
 
     /**
@@ -50,6 +56,9 @@ class FavoriteService
     public function destroyFavorite($product)
     {
         $user = User::findOrFail(auth()->user()->id);
+        if ($user->favoriteProducts()->where('product_id', $product->id)->doesntExist()) {
+            throw new Exception("You do not have permission to access this resource.");
+        }
         $user->favoriteProducts()->detach($product->id);
         $this->clearCacheGroup($this->groupe_key_cache);
     }
